+Created KeyProcessor Class (with two sub KeyProcessor classes) +Adjusted control framework in GameState and ControlManager +Fixed some GeneralMenu Issues -->Ability to change between states +Fixed the flow of input between the states and specifically towards chars +Adjusted the TestCharacter class +Adjusted the Controllable Interface
NOTES:
Almost ready to be merged with master once reviewed more.
As of now, I made the simple general menu with the ability to switch
between options. The only two that work are "EXIT" which closes out, and
"GameState" which is a test to switch to the game state. The game state
window displays the two characters with their own keyinputs and run
fine.

FUTURE:
-Going to update the character class more for using our own spritesheets
-Going to update the menu class.
-Going to change the name of some classes for clarity
-Going to document the methods and classes
-Might organize the classes into packages since the main package looks a
bit sloppy

Once framework is finished, Ill assign specific tasks to members to
update menu, characters, etc.

Basheer B.